@@ -48,11 +48,10 @@
     "object.entries": "^1.0.4",
     "react-test-renderer": "^16.8.6",
     "sinon": "^2.3.7",
-    "snyk": "^1.192.3",
+    "snyk": "^1.236.1",
     "webpack-cli": "^3.2.3"
   },
   "dependencies": {
-    "axios": "^0.19.0",
     "@babel/cli": "^7.2.3",
     "@babel/core": "^7.2.2",
     "@babel/plugin-transform-runtime": "^7.2.0",
@@ -60,21 +59,22 @@
     "@babel/preset-react": "^7.0.0",
     "@babel/preset-stage-0": "^7.0.0",
     "@babel/preset-stage-2": "^7.0.0",
+    "axios": "^0.19.0",
     "babel-loader": "^8.0.5",
     "css-loader": "^1.0.0",
     "identity-obj-proxy": "^3.0.0",
+    "lodash": "^4.17.15",
     "lodash.omit": "^4.5.0",
     "moment": "^2.20.1",
     "raven-js": "^3.24.1",
-    "react": "^16.8.6",
+    "react": "^16.10.2",
     "react-codemirror": "^1.0.0",
-    "react-dom": "^16.8.6",
-    "react-multistep": "^3.0.2",
+    "react-dom": "^16.10.2",
+    "react-multistep": "^3.6.13",
     "react-router": "^4.1.1",
     "react-router-dom": "^4.1.1",
     "react-sortable-table": "^1.3.0",
     "style-loader": "^0.23.1",
-    "lodash": ">=4.17.11",
     "webpack": "^4.0.0",
     "yarn": "^1.3.2"
   }
