@@ -3,7 +3,8 @@ import PropTypes from 'prop-types';
 
 class ErrorViewComponent extends React.Component {
     static propTypes = {
-        error: PropTypes.object.isRequired
+        error: PropTypes.object.isRequired,
+        newVersionComponent: PropTypes.object
     };
 
     /* expects axios error response in props.error */
@@ -21,6 +22,8 @@ class ErrorViewComponent extends React.Component {
         }
 
     }
+
+
     dictToList(dictObject){
         return <ul>
             {Object.keys(dictObject).map(key=>
@@ -42,7 +45,11 @@ class ErrorViewComponent extends React.Component {
         if (this.props.error.response) {
             // The request was made and the server responded with a status code
             // that falls out of the range of 2xx
-            return <p className="error-text">Server error {this.props.error.response.status}: {this.bestErrorString(this.props.error.response.data)}</p>
+            if(this.props.error.response.data.hasOwnProperty("nextAvailableVersion") && this.props.newVersionComponent!==null){
+                return <p className="error-text">This object already exists. {this.props.newVersionComponent}</p>
+            } else {
+                return <p className="error-text">Server error {this.props.error.response.status}: {this.bestErrorString(this.props.error.response.data)}</p>
+            }
         } else if (this.props.error.request) {
             // The request was made but no response was received
             // `error.request` is an instance of XMLHttpRequest in the browser and an instance of
